refactor: create reusable emoji component
