+++ app/containers/Emoji/index.ts
@@ -1,0 +1,4 @@
+import Emoji from './Emoji';
+import CustomEmoji from './CustomEmoji';
+
+export { Emoji, CustomEmoji };
