+++ app/containers/Emoji/interfaces.ts
@@ -1,0 +1,12 @@
+import { ImageStyle, StyleProp } from 'react-native';
+
+import { ICustomEmoji, IEmoji } from '../../definitions';
+
+export interface ICustomEmojiProps {
+	emoji: ICustomEmoji;
+	style: StyleProp<ImageStyle>;
+}
+
+export interface IEmojiProps {
+	emoji: IEmoji;
+}
