+++ app/containers/MessageComposer/components/Autocomplete/AutocompleteEmoji.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { View, Text } from 'react-native';
 
 import { IAutocompleteEmoji } from '../../interfaces';
-import { Emoji } from '../../../EmojiPicker/Emoji';
+import { Emoji } from '../../../Emoji';
 import { useStyle } from './styles';
 
 export const AutocompleteEmoji = ({ item }: { item: IAutocompleteEmoji }) => {
